+++ chartlets.py/tests/components/progress_test.py
@@ -13,8 +13,13 @@ class CircularProgressTest(make_base(CircularProgress)):
 
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
-            self.cls(color="success", value=10),
-            {"type": "CircularProgress", "color": "success", "value": 10},
+            self.cls(color="success", value=10, visible=False),
+            {
+                "type": "CircularProgress",
+                "color": "success",
+                "value": 10,
+                "visible": False,
+            },
         )
 
 
@@ -22,8 +27,13 @@ class CircularProgressWithLabelTest(make_base(CircularProgressWithLabel)):
 
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
-            self.cls(color="primary", value=12),
-            {"type": "CircularProgressWithLabel", "color": "primary", "value": 12},
+            self.cls(color="primary", value=12, visible=False),
+            {
+                "type": "CircularProgressWithLabel",
+                "color": "primary",
+                "value": 12,
+                "visible": False,
+            },
         )
 
 
@@ -31,8 +41,13 @@ class LinearProgressTest(make_base(LinearProgress)):
 
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
-            self.cls(color="success", value=40),
-            {"type": "LinearProgress", "color": "success", "value": 40},
+            self.cls(color="success", value=40, visible=False),
+            {
+                "type": "LinearProgress",
+                "color": "success",
+                "value": 40,
+                "visible": False,
+            },
         )
 
 
@@ -40,6 +55,11 @@ class LinearProgressWithLabelTest(make_base(LinearProgressWithLabel)):
 
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
-            self.cls(color="secondary", value=42),
-            {"type": "LinearProgressWithLabel", "color": "secondary", "value": 42},
+            self.cls(color="secondary", value=42, visible=False),
+            {
+                "type": "LinearProgressWithLabel",
+                "color": "secondary",
+                "value": 42,
+                "visible": False,
+            },
         )